+++ packages/preset-wind/src/variants/index.ts
@@ -6,4 +6,3 @@ export * from './experimental'
 export * from './media'
 export * from './misc'
 export * from './placeholder'
-export * from './variables'
